+++ src/WinGet.psm1
@@ -5,6 +5,7 @@
 $script:AcceptLicense = "AcceptLicense"
 $script:Force = "Force"
 $script:PackageSource = "WinGet"
+$script:AllVersions = "AllVersions"
 
 # Utility variables
 # Fast Package References are passed between cmdlets in the format of '<name>#<version>#<source>'